robust file loading
correctly handles:
- jp2a
- libcaca (img2txt)
- toilet
- files that omit color codes on some chars
- files that double-up color codes on a char

Currently, bdraw treats CR as any other character... meaning it will NOT
reproduce DOS-style line endings when re-saving the file!
The lines would be read like this:
...\r\n
However they would be saved like this:
...\r\e[0m\n
Perhaps this is a good reason to output escapes before color codes,
rather than after characters (like the other programs do), since then I
can preserve the line ending style for free.

I also might want to simply treat CRLF as a line ending. Treating CR as
a character creates a column of invisible characters at the right of the
image.

I think the ideal solution would be to treat them as line endings, but
include a global parameter that allows them to be treated as characters.